requirements.txt
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==1.0.3
opencv-python==4.12.0.88
piexif==1.1.3
pillow==11.3.0
ImageHash==4.3.2
numpy==2.0.0
gTTS==2.5.4
ExifRead==3.4.0
hachoir==3.3.0
ffmpeg-python==0.2.0
pillow-heif==1.1.0